# ts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmolVM SDK bridge",
    "summary": "A private local bridge for SmolVM SDK clients.",
    "version": "1"
  },
  "paths": {
...
    "/browser-sessions": {
      "post": {
        "summary": "Create Browser Session",
        "operationId": "createBrowserSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrowserSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browser-sessions/{session_id}": {
      "delete": {
        "summary": "Delete Browser Session",
        "operationId": "deleteBrowserSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes": {
      "get": {
        "summary": "List Sandboxes",
        "operationId": "listSandboxes",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SandboxResponse"
                  },
                  "type": "array",
                  "title": "Response Listsandboxes"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Sandbox",
        "operationId": "createSandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{sandbox_id}": {
      "get": {
        "summary": "Get Sandbox",
        "operationId": "getSandbox",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sandbox Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Sandbox",
        "operationId": "deleteSandbox",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sandbox Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{sandbox_id}/desktop": {
      "get": {
        "summary": "Get Sandbox Desktop",
        "operationId": "getSandboxDesktop",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sandbox Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesktopResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{sandbox_id}/cancel": {
      "post": {
        "summary": "Cancel Sandbox Operation",
        "description": "Stop an in-flight operation by deleting its session VM.",
        "operationId": "cancel_sandbox_operation_sandboxes__sandbox_id__cancel_post",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sandbox Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sandboxes/{sandbox_id}/exec": {
      "post": {
        "summary": "Exec Command",
        "operationId": "execCommand",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sandbox Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateBrowserSessionRequest": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_-]{0,62}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "headless",
              "live"
            ],
            "title": "Mode",
            "default": "headless"
          },
          "backend": {
            "type": "string",
            "enum": [
              "firecracker",
              "qemu",
              "libkrun",
              "auto"
            ],
            "title": "Backend",
            "default": "auto"
          },
          "profile_mode": {
            "type": "string",
            "enum": [
              "ephemeral",
              "persistent"
            ],
            "title": "Profile Mode",
            "default": "ephemeral"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-zA-Z0-9][a-zA-Z0-9_-]{0,62}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "timeout_minutes": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 1.0,
            "title": "Timeout Minutes",
            "default": 30
          },
          "viewport": {
            "$ref": "#/components/schemas/BrowserViewportRequest"
          },
          "record_video": {
            "type": "boolean",
            "title": "Record Video",
            "default": false
          },
          "allow_downloads": {
            "type": "boolean",
            "title": "Allow Downloads",
            "default": true
          },
          "memory": {
            "type": "integer",
            "maximum": 16384.0,
            "minimum": 512.0,
            "title": "Memory",
            "default": 2048
          },
          "disk_size": {
            "type": "integer",
            "maximum": 16384.0,
            "minimum": 2048.0,
            "title": "Disk Size",
            "default": 4096
          },
          "network": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OpenNetworkPolicy"
              },
              {
                "$ref": "#/components/schemas/OffNetworkPolicy"
              },
              {
                "$ref": "#/components/schemas/RestrictedNetworkPolicy"
              }
            ],
            "title": "Network",
            "discriminator": {
              "propertyName": "mode",
              "mapping": {
                "off": "#/components/schemas/OffNetworkPolicy",
                "open": "#/components/schemas/OpenNetworkPolicy",
                "restricted": "#/components/schemas/RestrictedNetworkPolicy"
              }
            }
          }
        },
        "type": "object",
        "title": "CreateBrowserSessionRequest",
        "description": "Create and boot a browser session owned by one SDK client."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Short explanation with a recovery action."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "The body returned for a handled API error."
      },
...
    }
  }
}